+++ src/Exceptions/ValueParserException.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * This file is a part of the webunit package.
+ *
+ * Read more at https://github.com/themichaelhall/webunit
+ */
+
+declare(strict_types=1);
+
+namespace MichaelHall\Webunit\Exceptions;
+
+use LogicException;
+
+/**
+ * Class representing an exception thrown when a value parse error occurred.
+ *
+ * @since 2.2.0
+ */
+class ValueParserException extends LogicException
+{
+}